net: sock: standardize the return value of `recv()'
The POSIX standard has defined the return value of `recv()' as
`ssize_t'. E/STACK attempts to be POSIX compatible, and should therefore
also define the return value of `recv()' as `ssize_t'.

+++ source/sockets/recv.c
@@ -85,7 +85,7 @@ ssize_t estack_recvfrom(int fd, void *buf, size_t length,
 	return -EINVALID;
 }
 
-int estack_recv(int fd, void *buf, size_t length, int flags)
+ssize_t estack_recv(int fd, void *buf, size_t length, int flags)
 {
 	UNUSED(flags);
 	return estack_recvfrom(fd, buf, length, flags, NULL, 0);
